Split test module in two.
Plus other minor tweaks
